move {get,set}_webhook_pr_state to webhook app (SOFTWARE-3325)
