+++ src/webhook_app.py
@@ -4,6 +4,7 @@ Application File
 import flask
 import flask.logging
 from flask import Flask, Response, request, render_template
+import glob
 import hmac
 import logging
 import os
@@ -80,6 +81,32 @@ def validate_webhook_signature(data, x_hub_signature):
         our_signature = "sha1=" + sha1
         return hmac.compare_digest(our_signature, x_hub_signature)
 
+def set_webhook_pr_state(num, sha, state):
+    prdir = "%s/%s" % (global_data.webhook_state_dir, num)
+    statefile = "%s/%s" % (prdir, sha)
+    os.makedirs(prdir, mode=0o755, exist_ok=True)
+    if isinstance(state, (tuple,list)):
+        state = "\n".join( x.replace("\n"," ") for x in map(str,state) )
+    with open(statefile, "w") as f:
+        print(state, file=f)
+
+def get_webhook_pr_state(sha, num='*'):
+    prdir = "%s/%s" % (global_data.webhook_state_dir, num)
+    statefile = "%s/%s" % (prdir, sha)
+    def path_check(fn): return re.search(r'/\d+/[a-f\d]{40}$', fn)
+    def pr_num(fn): return int(fn.rsplit('/', 2)[-2])
+    if num == '*':
+        filelist = glob.glob(statefile)
+        filelist = list(filter(path_check, filelist))
+        if len(filelist) == 0:
+            return None, None
+        # if there are multiple PRs with this sha, take the newest
+        statefile = max(filelist, key=pr_num)
+    if os.path.exists(statefile):
+        with open(statefile) as f:
+            return f.read().strip().split('\n'), pr_num(statefile)
+    else:
+        return None, None
 
 # already checked in automerge test script
 # might want to move the check here though, and pass in merge_sha
@@ -181,7 +208,7 @@ def status_hook():
         app.logger.info("Ignoring travis '%s' status hook" % ci_state)
         return Response("Not interested; CI state was '%s'" % ci_state)
 
-    pr_webhook_state, pull_num = global_data.get_webhook_pr_state(head_sha)
+    pr_webhook_state, pull_num = get_webhook_pr_state(head_sha)
     if pr_webhook_state is None or len(pr_webhook_state) != 4:
         app.logger.info("Got travis success status hook for commit %s;\n"
                 "not merging as No PR automerge info available" % head_sha)
@@ -268,7 +295,7 @@ def pull_request_hook():
     stdout, stderr, ret = runcmd(cmd, cwd=global_data.webhook_data_dir)
 
     webhook_state = (ret, base_sha, head_label, title)
-    global_data.set_webhook_pr_state(pull_num, head_sha, webhook_state)
+    set_webhook_pr_state(pull_num, head_sha, webhook_state)
 
     OK = "Yes" if ret == 0 else "No"
     Eligible = "Eligible!" if ret == 0 else "Not Eligible."
